Refactor tests, add coverage for building templates
